scheduler: Check InstanceManager when trying to reuse failed replicas
Longhorn #1304
