+++ controller/instance_manager_controller_test.go
@@ -2,13 +2,11 @@ package controller
 
 import (
 	"fmt"
-	"time"
 
 	"github.com/sirupsen/logrus"
 
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes/fake"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -56,41 +54,6 @@ func fakeInstanceManagerVersionUpdater(im *longhorn.InstanceManager) error {
 	im.Status.APIMinVersion = engineapi.CurrentInstanceManagerAPIVersion
 	im.Status.APIVersion = engineapi.CurrentInstanceManagerAPIVersion
 	return nil
-}
-
-func newInstanceManager(
-	name string,
-	imType types.InstanceManagerType,
-	currentState types.InstanceManagerState,
-	currentOwnerID, nodeID, ip string,
-	instances map[string]types.InstanceProcess,
-	isDeleting bool) *longhorn.InstanceManager {
-
-	im := &longhorn.InstanceManager{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: TestNamespace,
-			UID:       uuid.NewUUID(),
-			Labels:    types.GetInstanceManagerLabels(TestNode1, TestInstanceManagerImage, imType),
-		},
-		Spec: types.InstanceManagerSpec{
-			Image:  TestInstanceManagerImage,
-			NodeID: nodeID,
-			Type:   imType,
-		},
-		Status: types.InstanceManagerStatus{
-			OwnerID:      currentOwnerID,
-			CurrentState: currentState,
-			IP:           ip,
-			Instances:    instances,
-		},
-	}
-
-	if isDeleting {
-		now := metav1.NewTime(time.Now())
-		im.DeletionTimestamp = &now
-	}
-	return im
 }
 
 func newPod(phase v1.PodPhase, name, namespace, nodeID string) *v1.Pod {
