+++ scheduler/replica_scheduler.go
@@ -449,6 +449,15 @@ func (rcs *ReplicaScheduler) isFailedReplicaReusable(r *longhorn.Replica, v *lon
 		return false
 	}
 
+	im, err := rcs.ds.GetInstanceManagerByInstance(r)
+	if err != nil {
+		logrus.Errorf("failed to get instance manager when checking replica %v is reusable: %v", r.Name, err)
+		return false
+	}
+	if im.DeletionTimestamp != nil || im.Status.CurrentState != types.InstanceManagerStateRunning {
+		return false
+	}
+
 	return true
 }
